progress on grades script
